REFACTOR: app/auth -> pkg/auth
