+++ core/internal/app/flag/service/flag_service.go
@@ -2,12 +2,12 @@ package service
 
 import (
 	"context"
-	"core/internal/app/auth"
 	environmentrepo "core/internal/app/environment/repository"
 	flagmodel "core/internal/app/flag/model"
 	flagrepo "core/internal/app/flag/repository"
 	targetingrepo "core/internal/app/targeting/repository"
 	variationrepo "core/internal/app/variation/repository"
+	"core/internal/pkg/auth"
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"core/internal/pkg/srvenv"
