+++ core/internal/app/variation/service/variation_service.go
@@ -2,9 +2,9 @@ package service
 
 import (
 	"context"
-	"core/internal/app/auth"
 	variationmodel "core/internal/app/variation/model"
 	variationrepo "core/internal/app/variation/repository"
+	"core/internal/pkg/auth"
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"core/internal/pkg/srvenv"
