+++ core/internal/pkg/auth/auth_service.go
