+++ core/internal/pkg/auth/auth_util.go
